backend/requirements.txt
replicate
python-dotenv
flask_cors
openai
torch
flask
mangum